QMicroscope basic implementation
Added a QMicroscope widget for controlling/showing the state of a MicroscopeParameters object. Line number 505/506 and 598/599 must be commented to apply this change and is not active yet. It is currently not working seamlessly, especially the view of the dataframe, as this must be connected to a calibration file for proper implementation. A solution could be to reimplement the calibration file as a field in the MicroscopeParameter object. Or to create Calibration and CalibrationTable objects which can read calibrations from a table and distinguish them properly. This might be a big project, as it would also involve revisiting how the calibrations are applied to the MicroscopeParameters.